feat: build support --no-clean arg (umijs/father-next#12)
* feat: build noClean

* Apply suggestions from code review

+++ src/builder/index.ts
@@ -1,5 +1,5 @@
 import path from 'path';
-import { chokidar, rimraf } from '@umijs/utils';
+import { chokidar, rimraf, logger } from '@umijs/utils';
 import { IApi, IFatherConfig } from '../types';
 import bundle from './bundle';
 import bundless from './bundless';
@@ -29,6 +29,7 @@ interface IBuilderOpts {
   userConfig: IFatherConfig;
   cwd: string;
   pkg: IApi['pkg'];
+  clean?: boolean;
 }
 
 interface IWatchBuilderResult {
@@ -48,10 +49,13 @@ async function builder(
   const outputs = getProviderOutputs(configProviders);
   const watchers: chokidar.FSWatcher[] = [];
 
-  // clean output directories
-  outputs.forEach((output) => {
-    rimraf.sync(path.join(opts.cwd, output));
-  });
+  if (opts.clean !== false) {
+    // clean output directories
+    logger.info('Clean output directories');
+    outputs.forEach((output) => {
+      rimraf.sync(path.join(opts.cwd, output));
+    });
+  }
 
   if (!opts.watch && configProviders.bundle) {
     await bundle({
